Properly handle fclose return value when saving
In some instances, the write can be successful, but the final flush
to the filesystem fails because storage medium is out of space.

I had a case where I saved a file onto SMB-share that was out of space,
and notepad++ marked the buffer as not dirty, when in reality it failed
and the original file was overwritten by an empty one.